Move generic Condition/CondTree functions to the Types module.
